% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr.R
\name{CKiD_U25_combined_US}
\alias{CKiD_U25_combined_US}
\title{Cystatin C and serum Creatinine based eGFR by CKiD U25 equation (US units)}
\usage{
CKiD_U25_combined_US(cystatin, creat, age, sex, height, verbose = FALSE)
}
\arguments{
\item{cystatin}{numeric vector of Cystatin C levels in mg/dl}

\item{creat}{numeric vector of creatinine levels in mg/dl}

\item{age}{numeric vector of ages in years}

\item{sex}{character vector of sex ("F" = female, "M" = male)}

\item{height}{numeric vector of heights in cm}

\item{verbose}{a single boolean value. If true, return all the component
parts of the eGFR (cystatin, cr, combined), if false (defult), only return
the combined eGFR}
}
\value{
numeric vector of eGFR in ml/min/1.73m²
}
\description{
Wrapper function for estimating pediatric GFR based on data from the CKiD
study of children with CKD. Calculates a Cystatin C-based and Creatinine-
based eGFR then averages them. This has been shown to more faithfully
estimate measured GFR than either equation separately
}
\details{
Reference: Pierce CB, Muñoz A, Ng DK, Warady BA, Furth SL, Schwartz GJ.
Age- and sex-dependent clinical equations to estimate glomerular filtration
rates in children and young adults with chronic kidney disease. Kidney
International. 2021;99(4):948–956. doi:10.1016/j.kint.2020.10.047
}
\examples{
CKiD_U25_combined_US(cystatin = 1, creat = 0.7, age = 18, sex = "F",
                     height = 132, verbose = FALSE)
}
