% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr.R
\name{CKiD_U25_creatinine_US}
\alias{CKiD_U25_creatinine_US}
\title{Creatinine based eGFR by CKiD U25 equation (US units)}
\usage{
CKiD_U25_creatinine_US(creat, age, sex, height)
}
\arguments{
\item{creat}{numeric vector of creatinine levels in mg/dl}

\item{age}{numeric vector of ages in years}

\item{sex}{character vector of sex ("F" = female, "M" = male)}

\item{height}{numeric vector of heights in cm}
}
\value{
numeric vector of eGFR in ml/min/1.73m²
}
\description{
Reference: Pierce CB, Muñoz A, Ng DK, Warady BA, Furth SL, Schwartz GJ.
Age- and sex-dependent clinical equations to estimate glomerular filtration
rates in children and young adults with chronic kidney disease. Kidney
International. 2021;99(4):948–956. doi:10.1016/j.kint.2020.10.047
}
\examples{
CKiD_U25_creatinine_US(creat = 1, age = 12, sex = "F", height = 132)
}
